feat: signalement - comportement de la sélection et wording
