+++ src/store/index.ts
@@ -2,3 +2,4 @@ export { useUserStore } from "./useUserStore";
 export { useMapStore } from "./useMapStore";
 export { useCommunityStore } from "./useCommunityStore";
 export { useLocalStorageStore } from "./useLocalStorageStore";
+export { useReportStore } from "./useReportStore";
